Split SimOptions Settings page into SimOptions and Realism page
