+++ src/instruments/src/EFB/Settings/Settings.tsx
@@ -9,6 +9,7 @@ import { ScrollableContainer } from '../UtilComponents/ScrollableContainer';
 import { pathify, TabRoutes } from '../Utils/routing';
 import { AircraftOptionsPinProgramsPage } from './Pages/AircraftOptionsPinProgramsPage';
 import { SimOptionsPage } from './Pages/SimOptionsPage';
+import { RealismPage } from './Pages/RealismPage';
 import { AtsuAocPage } from './Pages/AtsuAocPage';
 import { AudioPage } from './Pages/AudioPage';
 import { FlyPadPage } from './Pages/FlyPadPage';
@@ -33,7 +34,7 @@ export const SelectionTabs = ({ tabs }: SelectionTabsProps) => (
             tabs.map((tab) => (
                 <Link
                     to={`settings/${pathify(tab.name)}`}
-                    className="flex justify-between items-center p-6 rounded-md border-2 border-transparent transition duration-100 bg-theme-secondary hover:border-theme-highlight"
+                    className="flex justify-between items-center p-6 bg-theme-secondary rounded-md border-2 border-transparent hover:border-theme-highlight transition duration-100"
                 >
                     <p className="text-2xl">{tab.name}</p>
                     <ChevronRight size={30} />
@@ -47,6 +48,7 @@ export const Settings = () => {
     const tabs: PageLink[] = [
         { name: 'Aircraft Options / Pin Programs', component: <AircraftOptionsPinProgramsPage /> },
         { name: 'Sim Options', component: <SimOptionsPage /> },
+        { name: 'Realism', component: <RealismPage /> },
         { name: 'ATSU / AOC', component: <AtsuAocPage /> },
         { name: 'Audio', component: <AudioPage /> },
         { name: 'flyPad', component: <FlyPadPage /> },
@@ -73,7 +75,7 @@ type SettingsPageProps = {
 export const SettingsPage: FC<SettingsPageProps> = ({ name, children }) => (
     <div>
         <Link to="/settings" className="inline-block mb-4">
-            <div className="flex flex-row items-center space-x-3 transition duration-100 hover:text-theme-highlight">
+            <div className="flex flex-row items-center space-x-3 hover:text-theme-highlight transition duration-100">
                 <ArrowLeft size={30} />
                 <h1 className="font-bold text-current">
                     Settings
@@ -82,7 +84,7 @@ export const SettingsPage: FC<SettingsPageProps> = ({ name, children }) => (
                 </h1>
             </div>
         </Link>
-        <div className="py-2 px-6 w-full rounded-lg border-2 h-efb border-theme-accent">
+        <div className="py-2 px-6 w-full h-efb rounded-lg border-2 border-theme-accent">
             <ScrollableContainer height={53}>
                 <div className="h-full divide-y-2 divide-theme-accent">
                     {children}
